requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy[asyncio]>=2.0.0
aiomysql>=0.2.0
alembic>=1.13.0
celery>=5.3.0
redis>=5.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
google-generativeai>=0.5.0
python-docx>=1.1.0
pypdf>=4.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
weasyprint>=61.0
python-pptx>=0.6.23
Jinja2>=3.1.0
httpx>=0.27.0
pdf2image>=1.17.0
lxml>=5.0.0
